Rename test_controller test_hub
