+++ tests/components/unifi/test_device_tracker.py
@@ -21,7 +21,7 @@ from homeassistant.core import HomeAssistant
 from homeassistant.helpers import entity_registry as er
 import homeassistant.util.dt as dt_util
 
-from .test_controller import ENTRY_CONFIG, setup_unifi_integration
+from .test_hub import ENTRY_CONFIG, setup_unifi_integration
 
 from tests.common import async_fire_time_changed
 from tests.test_util.aiohttp import AiohttpClientMocker
